Simplify journey data model
Combine monitored_journey and dated_journey,
so we get to operate mainly with the HFP data for now,
without depending on Jore data too much.
The idea of separating these would be that
  1) dated_journey is unique by (route, dir, oday, start), or better,
     (pattern_id, oday, start);
  2) monitored_journey is unique by the dated_journey reference AND the vehicle, i.e.,
     same dated_journey can be realized multiple times by different vehicles.
In this simplified model, we repeat dated_journey related attributes a bit.
Let's split the model later if needed.